+++ src/node-lib/curriculum-api-2023/queries/subjectListing/subjectListing.query.ts
@@ -1,20 +1,42 @@
+import { groupBy } from "lodash";
+
 import OakError from "../../../../errors/OakError";
 import { Sdk } from "../../sdk";
 
-import subjectListingSchema from "./subjectListing.schema";
+import subjectListingSchema, {
+  SubjectListingPageData,
+} from "./subjectListing.schema";
 
 const subjectListingQuery =
   (sdk: Sdk) => async (args: { keyStageSlug: string }) => {
     const res = await sdk.subjectListing(args);
     const [keyStageSubjects] = res.keyStageSubjects;
-    const subjectListing = {
-      ...keyStageSubjects,
-    };
 
     if (!keyStageSubjects) {
       throw new OakError({ code: "curriculum-api/not-found" });
     }
-    return subjectListingSchema.parse(subjectListing);
+
+    const { subjects } = keyStageSubjects;
+
+    const groupedSubjects = Object.values(
+      groupBy(
+        subjects,
+        (subject: SubjectListingPageData["subjects"][number]) =>
+          subject.subjectSlug
+      )
+    );
+
+    const subjectsWithProgrammeCount = groupedSubjects.map((subject) => {
+      return {
+        ...subject[0],
+        programmeCount: subject.length,
+      };
+    });
+
+    return subjectListingSchema.parse({
+      ...keyStageSubjects,
+      subjects: subjectsWithProgrammeCount,
+    });
   };
 
 export default subjectListingQuery;